feat(optimizations): move evaluate_expr_compile_time.rs to optimizations module
