+++ src/generator/mod.rs
@@ -1,7 +1,6 @@
 mod allocator;
 pub mod arm64;
 mod bingus;
-pub mod evaluate_expr_compile_time;
 pub mod function_validation;
 mod label;
 mod stack;
