+++ src/optimizer/mod.rs
@@ -1,1 +1,2 @@
+pub mod evaluate_expr_compile_time;
 pub mod fold_expr;
